+++ src/aui-form-builder/tests/unit/js/aui-form-builder-field-list-tests.js
@@ -28,6 +28,48 @@ YUI.add('aui-form-builder-field-list-tests', function(Y) {
             Y.Assert.isNotNull(Y.one('.form-builder-field'));
         },
 
+        'should add a field between fields in a list': function() {
+            var sentence = new Y.FormBuilderFieldSentence(),
+                sentence2 = new Y.FormBuilderFieldSentence();
+
+            this.createFieldList({ fields: [sentence, sentence2] });
+
+            Y.all('.form-builder-field-list-add-button').item(1).simulate('mouseover');
+            Y.all('.form-builder-field-list-add-button-circle').item(1).simulate('click');
+
+            this._fieldList.addField(new Y.FormBuilderFieldSentence({title: 'between'}));
+
+            Y.Assert.areEqual(Y.all('.form-field-title').item(1).get('innerHTML'), 'between');
+        },
+
+        'should show/hide add button between fields on mouseover/mouseleave': function() {
+            var addButton,
+                sentence = new Y.FormBuilderFieldSentence(),
+                sentence2 = new Y.FormBuilderFieldSentence();
+
+            this.createFieldList({ fields: [sentence, sentence2] });
+
+            addButton = Y.all('.form-builder-field-list-add-button').item(1);
+
+            Y.Assert.isFalse(addButton.hasClass('form-builder-field-list-add-button-visible'));
+
+            addButton.simulate('mouseover');
+            Y.Assert.isTrue(addButton.hasClass('form-builder-field-list-add-button-visible'));
+
+            addButton.simulate('mouseout');
+            Y.Assert.isFalse(addButton.hasClass('form-builder-field-list-add-button-visible'));
+
+            addButton = Y.all('.form-builder-field-list-add-button').item(2);
+
+            Y.Assert.isTrue(addButton.hasClass('form-builder-field-list-add-button-visible'));
+
+            addButton.simulate('mouseover');
+            Y.Assert.isTrue(addButton.hasClass('form-builder-field-list-add-button-visible'));
+
+            addButton.simulate('mouseout');
+            Y.Assert.isTrue(addButton.hasClass('form-builder-field-list-add-button-visible'));
+        },
+
         'should remove a field from the field list': function() {
             var sentence = new Y.FormBuilderFieldSentence();